Refactor link-checker utility and update routing paths
- Introduced a new link-checker utility with enhanced link validation and extraction capabilities, including support for dynamic routes and patterns.
- Removed the old `link-checker.ts` file to streamline the codebase.
- Updated routing paths in `ResearchForm.tsx` and `GenericVariableAddButton.tsx` to simplify navigation by removing redundant segments.
- Improved integration tests for the link-checker functionality.
